backend/app/pyproject.toml
[tool.poetry]
name = "w-app-backend"
version = "0.1.0"
description = "A simple app"
authors = ["Urich Oliveira <oliveira.urich@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
dynaconf = "^3.1.2"
uvicorn = "^0.13.3"
SQLAlchemy = "^1.3.23"
arrow = "^0.17.0"
passlib = "^1.7.4"
python-jose = "^3.2.0"
bcrypt = "^3.2.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
